improve dev experience

+++ gopkg/conn/pg/pool.go
@@ -12,8 +12,8 @@ func New(ctx context.Context, cfg Config) (*pgxpool.Pool, error) {
 		return nil, err
 	}
 
-	poolCfg.MaxConns = int32(cfg.MaxOpenConns)
-	poolCfg.MinConns = int32(cfg.MaxIdleConns)
+	poolCfg.MaxConns = cfg.MaxOpenConns
+	poolCfg.MinConns = cfg.MaxIdleConns
 	poolCfg.MaxConnLifetime = cfg.ConnMaxLifetime
 
 	return pgxpool.NewWithConfig(ctx, poolCfg)
